modules/image_io: support non-integer scales

+++ modules/image_io/image_gif.h
@@ -9,7 +9,7 @@
 iPoint image_size_gif(const std::string & file);
 
 // load the whole image
-Image image_load_gif(const std::string & file, const int scale=1);
+Image image_load_gif(const std::string & file, const double scale=1);
 
 // save the whole image
 void image_save_gif(const Image & im, const std::string & file,
